Add style guide for ColumnsTable and ToolbarCOnfig components
