+++ src/components/ColumnsTable/index.js
@@ -1,0 +1,3 @@
+import ColumnsTable from './ColumnsTable.component';
+
+export default ColumnsTable;
